requirements.txt
scrapy
scrapy-playwright
playwright
gitdb==4.0.11
GitPython==3.1.41
psycopg2==2.9.10
setuptools==69.0.3
smmap==5.0.1
